refacyor: parser in different files for readability
